+++ backend/tests/ui_handling/conftest.py
@@ -8,12 +8,7 @@ def app():
     app.config.update({
         "TESTING": True,
     })
-
-    # other setup can go here
-
     yield app
-
-    # clean up / reset resources here
 
 
 @pytest.fixture()
