+++ backend/tests/ui_handling/test_get_endpoints.py
